polish

+++ init.php
@@ -47,8 +47,14 @@ if ($page == "index") {
 	}
 }
 
-if ($page == "inscription") {
+if ($page == "connexion") {
 	if (ft_logged()) {
+		if (isset($_GET['deconnexion'])) {
+			unset($_SESSION['id']);
+		} else if (isset($_GET['desactivation'])) {
+			mysqli_query($sql, "UPDATE users SET desactivated = 1 WHERE id = ".$_SESSION['id']);
+			unset($_SESSION['id']);
+		}
 		header("Location: index.php");
 		exit;
 	}
@@ -75,19 +81,6 @@ if ($page == "inscription") {
 			header("Location: connexion.php");
 			exit;
 		}
-	}
-}
-
-if ($page == "connexion") {
-	if (ft_logged()) {
-		if (isset($_GET['deconnexion'])) {
-			unset($_SESSION['id']);
-		} else if (isset($_GET['desactivation'])) {
-			mysqli_query($sql, "UPDATE users SET desactivated = 1 WHERE id = ".$_SESSION['id']);
-			unset($_SESSION['id']);
-		}
-		header("Location: index.php");
-		exit;
 	}
 	if (isset($_POST['connexion'])) {
 		$email = $_POST['email'];
@@ -133,18 +126,18 @@ if ($page == "valider") {
 	}
 }
 
-// if ($page == "admin") {
-// 	if (!ft_admin()) {
-// 		header("Location: index.php");
-// 		exit;
-// 	}
-// }
+if ($page == "admin") {
+	if (!ft_admin()) {
+		header("Location: index.php");
+		exit;
+	}
+}
 
 if ($page == "admin_users") {
-	// if (!ft_admin()) {
-	// 	header("Location: index.php");
-	// 	exit;
-	// }
+	if (!ft_admin()) {
+		header("Location: index.php");
+		exit;
+	}
 	if (isset($_GET['del_id'])) {
 		$del_id = ceil($_GET['del_id']);
 		mysqli_query($sql, "DELETE FROM users WHERE id = ".$del_id);
@@ -162,10 +155,10 @@ if ($page == "admin_users") {
 }
 
 if ($page == "admin_articles") {
-	// if (!ft_admin()) {
-	// 	header("Location: index.php");
-	// 	exit;
-	// }
+	if (!ft_admin()) {
+		header("Location: index.php");
+		exit;
+	}
 	$id_categorie = ceil(@$_GET['id']);
 	$retour = mysqli_query($sql, "SELECT * FROM categories WHERE id = ".$id_categorie);
 	if (mysqli_num_rows($retour) == 0) {
@@ -195,10 +188,10 @@ if ($page == "admin_articles") {
 }
 
 if ($page == "admin_categories") {
-	// if (!ft_admin()) {
-	// 	header("Location: index.php");
-	// 	exit;
-	// }
+	if (!ft_admin()) {
+		header("Location: index.php");
+		exit;
+	}
 	if (isset($_GET['del_id'])) {
 		$del_id = ceil($_GET['del_id']);
 		if (!$articles = mysqli_query($sql, "SELECT * FROM articles WHERE gamme = ".$del_id))
@@ -248,10 +241,10 @@ if ($page == "admin_categories") {
 }
 
 if ($page == "admin_orders") {
-	// if (!ft_admin()) {
-	// 	header("Location: index.php");
-	// 	exit;
-	// }
+	if (!ft_admin()) {
+		header("Location: index.php");
+		exit;
+	}
 	if (isset($_GET['valid_id'])) {
 		$valid_id = ceil($_GET['valid_id']);
 		mysqli_query($sql, "UPDATE panier SET finished = finished + 1 WHERE id = ".$valid_id);
@@ -264,12 +257,12 @@ if ($page == "admin_orders") {
 }
 
 if ($page == "admin_order") {
-	// if (!ft_admin()) {
-	// 	header("Location: index.php");
-	// 	exit;
-	// }
+	if (!ft_admin()) {
+		header("Location: index.php");
+		exit;
+	}
 	$id_commande = ceil(@$_GET['id']);
-	$retour = mysqli_query($sql, "SELECT * FROM shop WHERE id = ".$id_commande);
+	$retour = mysqli_query($sql, "SELECT * FROM panier WHERE id = ".$id_commande);
 	if (mysqli_num_rows($retour) == 0) {
 		header("Location: admin_orders.php");
 		exit;
